* Transfer code from speech-diffusion-streaming-vc
* Facilitate the inference
* Support non-overlap global prompt; add p_drop_bn_ctx
* Support short-len BN data
* Fix: prompt_bn use max_prompt_len
* Transfer from speech-diffusion-notoken-dev1
* Add valid set; fix inference
* Support consistency distillation
* Support inference of consistency model; Fix hanging bug
* Support equal-length prompt_bn
* Fix a bug in CD; modify the default config of CD
* Support pseudo-huber loss and logit-normal sampling
* Fix a bug of psuedo-huber loss
* Update criterions format
* Fix: descretize Consistency distillation
* Support adversarial distill
* Fix odesolver for consistency distill; support student/teacher CFG during CD
* Update the config of adversarial distill; add diffusion loss
* Update adv distill with many features; add ema to CD
* Support adversarial consistency model
* Support formal editing inference
* Support separate CFG inference
* Support guidance embedding in LCM
* Fix SSIM loss
* add config
* Support ACD
* Merge branch 'speech-dit-dev-icm2' into speechdit
* update editing script
* Modify scripts
* Support speed editing mode
* Add L1 yaml; fix rectified flow
* Fix min lr
* dit-edit support en/zh-en
* Modify the rule for determining the mask length for editing; use AA instead of A
* [feat] dit support multi-lingual
* [feat] DiT 8-lang version for both training and inference
* [fix] fix merge error
* [reorg] remvove useless reliance
* [feat] add yaml config
* [fix] fix reliance
* [fix] fix length bug
* [fix] remove adaptive;remove phone alignments;edit infer script
* [feat] support cla-concat mode
* Merge branch 'speechdit' of code.byted.org:seed/samantha into speechdit
* fix: fix error
* [fix] fix bugs in concat infer
* [fix] remove useless codes
* dit edit preprocess merge
* Finish the evaluation of TTS, Editing and CLA
* Merge branch 'speechdit'
...

https://code.byted.org/seed/samantha/merge_requests/1156

+++ samantha/models/ECAPA_TDNN_bias.py
@@ -197,6 +197,7 @@ class ECAPA_TDNN_GN(nn.Module):
         sg = torch.sqrt((torch.sum((x**2) * w, dim=2) - mu**2).clamp(min=1e-4))
 
         x = torch.cat((mu, sg), 1)
+
         x = self.bn5(x)
-        x = self.fc6(x)
+        x = self.fc6(x).float()
         return x
